+++ spacy/cli/_app.py
@@ -1,31 +1,31 @@
+from typing import Optional
 import typer
-from wasabi import msg
+from typer.main import get_command
 
 
-def Arg(*args, help=None, **kwargs):
+COMMAND = "python -m spacy"
+NAME = "spacy"
+HELP = """spaCy Command-line Interface
+
+DOCS: https://spacy.io/api/cli
+"""
+
+
+app = typer.Typer(name=NAME, help=HELP)
+
+
+def Arg(*args, help: Optional[str] = None, **kwargs) -> typer.Argument:
+    """Wrapper for Typer's annotation to keep it short and set defaults."""
     # Filter out help for now until it's officially supported
     return typer.Argument(*args, **kwargs)
 
 
-def Opt(*args, **kwargs):
+def Opt(*args, **kwargs) -> typer.Option:
+    """Wrapper for Typer's annotation to keep it short and set defaults."""
     return typer.Option(*args, show_default=True, **kwargs)
 
 
-app = typer.Typer(
-    name="spacy",
-    help="""spaCy Command-line Interface
-
-
-DOCS: https://spacy.io/api/cli
-""",
-)
-
-
-@app.command("link", no_args_is_help=True, deprecated=True, hidden=True)
-def link(*args, **kwargs):
-    """As of spaCy v3.0, model symlinks are deprecated. You can load models
-    using their full names or from a directory path."""
-    msg.warn(
-        "As of spaCy v3.0, model symlinks are deprecated. You can load models "
-        "using their full names or from a directory path."
-    )
+def setup_cli() -> None:
+    # Ensure that the help messages always display the correct prompt
+    command = get_command(app)
+    command(prog_name=COMMAND)
